Re-did the internal structure of the KeyPadView's look.
Instead of using a single Grid with a bunch of columns and rows, I'm now
using Grid's within Grids. This lets me group related buttons with-in a
sub-grid, and position them with relative to each-other instead of
positioning them based on their final location. It makes positioning and
re-positioning a bit easier. Actually, it makes re-positioning a lot
easier.